app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-docx==1.1.2
python-multipart==0.0.9
Pillow
requests==2.32.3
lxml
